add configGit configurar o name email

+++ configGit.sh
@@ -1,5 +1,14 @@
 UserWindows=$(powershell.exe '$env:UserName' | tr -d '\r')
-cp /mnt/c/Users/$UserWindows/.gitconfig .gitconfig
+if [ -f "/mnt/c/Users/$UserWindows/.gitconfig" ]; then
+    cp /mnt/c/Users/$UserWindows/.gitconfig .gitconfig
+else
+    echo name:?
+    read username
+    echo email?
+    read useremail
+    git config --global user.name $username
+    git config --global user.email $useremail
+fi
 
 if [ -f "/mnt/c/Program Files/Git/mingw64/bin/git-credential-manager.exe" ]; then
     git config --global credential.helper "/mnt/c/Program\ Files/Git/mingw64/bin/git-credential-manager.exe"
